requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.6.0
anthropic>=0.39.0
pyyaml>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
email-validator>=2.1.0
alembic>=1.13.0
tenacity>=8.2.0
